Version 0.0.2
This is the biggest update yet.
This update allows the creation of levels extremely easily using an interface called level. The game class is reduced in size to the bare minimum. Saving the map with (Ctrl + s) now opens a dialog to save the map txt where ever you want. Alpha quest will now programically check if it is running as an exe and will look at the assets folder in eclipse and as an install out side of eclipse effortlessly. If you want instructions on making an exe/installer check out the guide to making an installer on [Creating an EXE and Installer](https://github.com/mikezamora1998/AlphaQuest/wiki/Creating-an-EXE-and-Installer)

Co-Authored-By: btcarlsen <btcarlsen@users.noreply.github.com>
Co-Authored-By: dlichliter <dlichliter@users.noreply.github.com>